+++ src/pages/todos/TodoCreatePage.tsx
@@ -9,23 +9,14 @@ import {
 } from '@/schemas/todoCreateSchema';
 import { validateTextMaxLength } from '@/utils/validators';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 
 import { Controller, useForm } from 'react-hook-form';
 import AppButton from '@/components/common/AppButton';
-import ToggleSwitch from '@/components/common/ToggleSwitch';
-import { Check, ChevronDown } from 'lucide-react';
-import { cn } from '@/lib/utils';
-
-const REPEAT_CYCLE_OPTIONS = ['매주', '매월'] as const;
-type RepeatCycle = (typeof REPEAT_CYCLE_OPTIONS)[number];
+import RepeatSection from './components/RepeatSection';
 
 const TodoCreatePage = () => {
   const [inputErrorMessage, setInputErrorMessage] = useState('');
-  const [isRepeatEnabled, setIsRepeatEnabled] = useState(false);
-  const [repeatCycle, setRepeatCycle] = useState<RepeatCycle>('매주');
-  const [isRepeatMenuOpen, setIsRepeatMenuOpen] = useState(false);
-  const repeatCycleDropdownRef = useRef<HTMLDivElement>(null);
 
   const {
     control,
@@ -42,33 +33,6 @@ const TodoCreatePage = () => {
   const onSubmit = (data: TodoCreateValues) => {
     console.log('할 일 등록: ', data);
   };
-
-  useEffect(() => {
-    if (!isRepeatMenuOpen) return;
-
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        repeatCycleDropdownRef.current &&
-        !repeatCycleDropdownRef.current.contains(event.target as Node)
-      ) {
-        setIsRepeatMenuOpen(false);
-      }
-    };
-
-    const handleEscape = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        setIsRepeatMenuOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    document.addEventListener('keydown', handleEscape);
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-      document.removeEventListener('keydown', handleEscape);
-    };
-  }, [isRepeatMenuOpen]);
 
   return (
     <FormPageLayout
@@ -119,6 +83,7 @@ const TodoCreatePage = () => {
         <p className="text-sm font-semibold text-[#BCBCBC]">상세 설정</p>
 
         <div className="flex flex-col gap-[30px] pt-6">
+          {/* 마감기한 */}
           <section className="space-y-4">
             <p className="text-base font-semibold">마감기한</p>
 
@@ -128,6 +93,7 @@ const TodoCreatePage = () => {
             </AppButton>
           </section>
 
+          {/* 담당자 */}
           <section className="space-y-4">
             <p className="text-base font-semibold">담당자</p>
 
@@ -151,91 +117,10 @@ const TodoCreatePage = () => {
             </div>
           </section>
 
-          <section className="space-y-4 text-base font-semibold">
-            <div className="flex items-center justify-between">
-              <p>반복</p>
+          {/* 반복 */}
+          <RepeatSection />
 
-              <ToggleSwitch
-                checked={isRepeatEnabled}
-                onCheckedChange={(checked) => {
-                  setIsRepeatEnabled(checked);
-
-                  if (!checked) {
-                    setIsRepeatMenuOpen(false);
-                  }
-                }}
-                ariaLabel="반복 설정 토글"
-              />
-            </div>
-
-            {isRepeatEnabled && (
-              <div className="flex flex-col gap-[15px]">
-                <div
-                  className="relative flex justify-end"
-                  ref={repeatCycleDropdownRef}
-                >
-                  <button
-                    type="button"
-                    className="flex items-center gap-1"
-                    onClick={() => setIsRepeatMenuOpen((prevOpen) => !prevOpen)}
-                    aria-haspopup="menu"
-                    aria-expanded={isRepeatMenuOpen}
-                  >
-                    <span>{repeatCycle}</span>
-                    <ChevronDown
-                      className={cn(
-                        'h-6 w-6 transition-transform',
-                        isRepeatMenuOpen && 'rotate-180'
-                      )}
-                    />
-                  </button>
-
-                  {isRepeatMenuOpen && (
-                    <ul
-                      role="menu"
-                      className="absolute top-[calc(100%+6px)] right-0 z-10 w-[77px] rounded-md border border-[#898887] bg-[#696867] p-1 text-xs font-medium text-white shadow-md"
-                    >
-                      {REPEAT_CYCLE_OPTIONS.map((option) => (
-                        <li key={option}>
-                          <button
-                            type="button"
-                            role="menuitemradio"
-                            aria-checked={repeatCycle === option}
-                            className="flex w-full items-center gap-1 rounded-md px-2 py-1.5 text-left hover:bg-[#525150]"
-                            onClick={() => {
-                              setRepeatCycle(option);
-                              setIsRepeatMenuOpen(false);
-                            }}
-                          >
-                            <div className="flex gap-1">
-                              <span>
-                                {repeatCycle === option ? (
-                                  <Check className="h-4 w-4" />
-                                ) : (
-                                  <div className="h-4 w-4" />
-                                )}
-                              </span>
-                              <span className="text-xs font-medium">
-                                {option}
-                              </span>
-                            </div>
-                          </button>
-                        </li>
-                      ))}
-                    </ul>
-                  )}
-                </div>
-
-                {repeatCycle === '매주' && <div>월 ~ 금</div>}
-
-                <div className="flex flex-col gap-[15px]">
-                  <div>시작 날짜</div>
-                  <div>종료 날짜</div>
-                </div>
-              </div>
-            )}
-          </section>
-
+          {/* 메모 */}
           <section className="space-y-4">
             <p className="text-base font-semibold">메모</p>
